+++ sphaera/diffeq/_impl/tsit5.py
@@ -52,7 +52,7 @@ def _optimal_step_size(last_step, mean_error_ratio, safety=0.9, ifactor=10.0, df
     if mean_error_ratio == 0:
         return last_step * ifactor
     if mean_error_ratio < 1:
-        dfactor = _convert_to_tensor(1, dtype=torch.float64, device=mean_error_ratio.device)
+        dfactor = _convert_to_tensor(1, dtype=torch.float32, device=mean_error_ratio.device)
     error_ratio = torch.sqrt(mean_error_ratio).type_as(last_step)
     exponent = torch.tensor(1 / order).type_as(last_step)
     factor = torch.max(1 / ifactor, torch.min(error_ratio**exponent / safety, 1 / dfactor))
@@ -77,9 +77,9 @@ class Tsit5Solver(AdaptiveStepsizeODESolver):
         self.rtol = rtol
         self.atol = atol
         self.first_step = first_step
-        self.safety = _convert_to_tensor(safety, dtype=torch.float64, device=y0[0].device)
-        self.ifactor = _convert_to_tensor(ifactor, dtype=torch.float64, device=y0[0].device)
-        self.dfactor = _convert_to_tensor(dfactor, dtype=torch.float64, device=y0[0].device)
+        self.safety = _convert_to_tensor(safety, dtype=torch.float32, device=y0[0].device)
+        self.ifactor = _convert_to_tensor(ifactor, dtype=torch.float32, device=y0[0].device)
+        self.dfactor = _convert_to_tensor(dfactor, dtype=torch.float32, device=y0[0].device)
         self.max_num_steps = _convert_to_tensor(max_num_steps, dtype=torch.int32, device=y0[0].device)
 
     def before_integrate(self, t):
